+++ src/libs/core/exception/warning.hpp
@@ -4,6 +4,7 @@
 #include <iostream>
 #include <ostream>
 #include <vector>
+#include <cinttypes>
 
 class Warning
 {
@@ -14,7 +15,7 @@ public:
         Argument
     };
 
-    Warning(Type _type, const std::string& _message, int _line, int _column);
+    Warning(Type _type, const std::string& _message, int64_t _line, int64_t _column);
 
     const char* what() const noexcept;
     Type getType() const noexcept;
@@ -22,14 +23,14 @@ public:
     void print(std::ostream& os = std::cerr) const;
 
 
-    static Warning GeneralWarning(const std::string& message, int line = -1, int column = -1);
-    static Warning ArgumentWarning(const std::string& message, int line = -1, int column = -1);
+    static Warning GeneralWarning(const std::string& message, int64_t line = -1, int64_t column = -1);
+    static Warning ArgumentWarning(const std::string& message, int64_t line = -1, int64_t column = -1);
 
 private:
     Type type;
     std::string message;
-    int line;
-    int column;
+    int64_t line;
+    int64_t column;
 
     std::string typeToString() const;
 };
